requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pandas
numpy
requests
python-dotenv
openai
